WIP. Clippy lints and project tweaks.
WIP: Anything else?

Fix a bunch of clippy lints (both new and those that somehow got missed
from earlier releases).

Change the optimisation level of Cargo's dev profile to 0, but leave
test at 2. This hopefully means that tests aren't unbearably slow, but
things like check and clippy are also faster.

+++ src/model/cuda.rs
@@ -445,7 +445,7 @@ impl<'a> SkyModellerCuda<'a> {
                 i_unflagged_tile += 1;
                 continue;
             }
-            tile_index_to_unflagged_tile_index_map.push(i_unflagged_tile as i32);
+            tile_index_to_unflagged_tile_index_map.push(i_unflagged_tile);
             i_unflagged_tile += 1;
         }
         let d_tile_index_to_unflagged_tile_index_map =
